+++ cudax/test/copy_bytes/mdspan_d2h_h2d_relaxed.cu
@@ -72,7 +72,7 @@ TEST_CASE("copy_bytes layout_stride_relaxed, sliding window", "[copy_bytes][rela
   {
     for (int j = 0; j < N; ++j)
     {
-      expected[i * N + j] = i + j;
+      expected[static_cast<std::size_t>(i) * N + j] = i + j;
     }
   }
 
@@ -118,7 +118,7 @@ TEST_CASE("copy_bytes layout_stride_relaxed, 2D column-reversed", "[copy_bytes][
   {
     for (int j = 0; j < N; ++j)
     {
-      expected[i * N + j] = i * N + (N - 1 - j);
+      expected[static_cast<std::size_t>(i) * N + j] = static_cast<std::size_t>(i) * N + (N - 1 - j);
     }
   }
   using extents_t = cuda::std::extents<int, M, N>;
